Install graphql-subscriptions, subscriptions-transport-ws, apollo-link-ws and apollo-utilities modules. Wrapp app with createServer and add subscriptionServer. Uninstall socket module.
