+++ client/package.json
@@ -8,6 +8,8 @@
     "apollo-link": "^1.2.1",
     "apollo-link-context": "^1.0.7",
     "apollo-link-http": "^1.5.3",
+    "apollo-link-ws": "^1.0.7",
+    "apollo-utilities": "^1.0.11",
     "axios": "^0.15.3",
     "babel-polyfill": "^6.26.0",
     "classnames": "^2.2.5",
@@ -47,6 +49,7 @@
     "semantic-ui-transition": "^2.2.10",
     "semantic-ui-visibility": "^2.2.10",
     "shortid": "^2.2.8",
+    "subscriptions-transport-ws": "^0.9.7",
     "validator": "^7.0.0",
     "whatwg-fetch": "^2.0.3"
   },
