+++ server/package.json
@@ -35,6 +35,7 @@
     "express-session": "^1.15.6",
     "express-sessions": "^1.0.6",
     "graphql": "^0.13.1",
+    "graphql-subscriptions": "^0.5.8",
     "graphql-tools": "^2.21.0",
     "jsonwebtoken": "^8.1.1",
     "lodash": "^4.17.4",
@@ -46,7 +47,8 @@
     "passport-local": "^1.0.0",
     "pg": "^7.4.1",
     "sequelize": "^4.36.0",
-    "socket.io": "^2.0.4"
+    "socket.io": "^2.0.4",
+    "subscriptions-transport-ws": "^0.9.7"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
